CLKernelExecutor: throw an exception when kernel execution fails
Throw an exception when the kernel execution failes. That way the calling
code can react to the exception and recover from the failure.

I need this change to catch the OpenCLException
CL_MEM_OBJECT_ALLOCATION_FAILURE, which tells me, that the GPU runs out
of memory. I then free some memory and rerun the kernel.